requirements.txt
torch
torchvision
diffusers
transformers
accelerate
safetensors
datasets
pillow
tqdm
# Optional: wandb
peft